requirements.txt
PyYAML==6.0.1
requests==2.31.0
prometheus-client==0.19.0
redis>=4.5.0,<5.0.0
